LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/timeline-scope.html

<!DOCTYPE html>
<title>Behavior of the timeline-scope property</title>
<link rel="help" src="https://github.com/w3c/csswg-drafts/issues/7759">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>

<main id=main></main>
<script>
  async function inflate(t, template) {
    t.add_cleanup(() => main.replaceChildren());
    return runAndWaitForFrameUpdate(() => {
       main.append(template.content.cloneNode(true));
    });
  }

  async function scrollTop(e, value) {
    e.scrollTop = value;
    await waitForNextFrame();
  }
</script>
<style>
  @keyframes anim {
    from { width: 0px; }
    to { width: 200px; }
  }

  .scroller {
    overflow-y: hidden;
    width: 200px;
    height: 200px;
  }
  .scroller > .content {
    margin: 400px 0px;
    width: 100px;
    height: 100px;
    background-color: green;
  }
  .target {
    background-color: coral;
    width: 0px;
    animation: anim auto linear;
    animation-timeline: --t1;
  }
  .timeline {
    scroll-timeline-name: --t1;
  }
  .scope {
    timeline-scope: --t1;
  }

</style>

<template id=deferred_timeline_appearing>
  <div class=container>
    <div class=target>Test</div>
    <div class="scroller timeline">
      <div class=content></div>
    </div>
  </div>
</template>
<script>
  promise_test(async (t) => {
    await inflate(t, deferred_timeline_appearing);
    let container = main.querySelector('.container');
    let scroller = main.querySelector('.scroller');
    let target = main.querySelector('.target');

    await scrollTop(scroller, 350); // 50%

    // Not attached to any timeline initially.
    assert_equals(getComputedStyle(target).width, '0px');

    // Add the deferred timeline.
    container.classList.add('scope');
    await waitForNextFrame();
    assert_equals(getComputedStyle(target).width, '100px'); // 0px => 200px, 50%

    // Remove the deferred timeline.
    container.classList.remove('scope');
    await waitForNextFrame();
    assert_equals(getComputedStyle(target).width, '0px');
  }, 'A deferred timeline appearing dynamically in the ancestor chain');
</script>
