LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/timeline-scope-expected.txt


FAIL Descendant can attach to deferred timeline promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'anim.ready')"
PASS Deferred timeline with no attachments
FAIL Inner timeline does not interfere with outer timeline assert_equals: expected "100px" but got "0px"
PASS Deferred timeline with two attachments
FAIL Dynamically re-attaching assert_equals: expected "50px" but got "100px"
FAIL Dynamically detaching assert_equals: expected "0px" but got "50px"
FAIL Removing/inserting element with attaching timeline assert_equals: expected "0px" but got "100px"
PASS Ancestor attached element becoming display:none/block
FAIL A deferred timeline appearing dynamically in the ancestor chain assert_equals: expected "0px" but got "100px"
FAIL Animations prefer non-deferred timelines assert_equals: expected "150px" but got "0px"
